fix: move explain methods to clickhousetelemetrystore
