+++ pkg/telemetrystore/clickhousetelemetrystore/explain_statement_test.go
@@ -1,4 +1,4 @@
-package telemetrystore
+package clickhousetelemetrystore
 
 import "testing"
